Refactor supabase client usage
