+++ src/services/retell/retellAgentSync.ts
@@ -1,5 +1,5 @@
 
-import { supabase } from '@/lib/supabase';
+import { supabase } from '@/integrations/supabase/client';
 import { retellApiDebugger } from './retellApiDebugger';
 
 export interface AgentSyncResult {
